Moves "__main__" dataset extraction code
- Moves extraction code from "__main__.py" to "extract_dataset.py"
